Modified html templates, added css

+++ app/blueprints/profile_bp.py
@@ -1,3 +1,4 @@
+from app.db_interactors.drink_db_inter import DrinkDbInter
 from app.db_interactors.img_db_inter import ImgDbInter
 from app.db_interactors.user_db_inter import UserDbInter
 from app.interactors.img_inter import ImgInter
@@ -16,16 +17,14 @@ profile_bp = Blueprint('profile', __name__)
 
 @profile_bp.route('/v1/profile/<user_id>', methods=['GET'])
 def display_profile(user_id):
+    msg = None
     user = UserDbInter().get_user(user_id)
     img = ImgInter().get_img_path(user)
-    return render_template('profile.html', title=user.nick, user=user, img=img)
-
-
-@login_required
-@profile_bp.route('/v1/profile/manage/<user_id>')
-def manage_profile(user_id):
-    user = UserDbInter().get_user(user_id)
-    return render_template('manage_profile.html', title=user.nick)
+    drinks = DrinkDbInter().search_by_user(user_id)
+    if not drinks:
+        msg = 'This user did not add any drinks yet.'
+    return render_template('profile.html', title=user.nick, user=user, img=img,
+                           drinks=drinks, msg=msg)
 
 
 @login_required
@@ -38,13 +37,14 @@ def change_password():
         e = Validators().validate_pass(d['newpass'], d['confirmnew'])
         if check_password_hash(current_user.password_hash, d['oldpass']):
             if e:
-                flash(e)
+                flash(e, category='error')
             else:
                 UserDbInter().update_password(d['newpass'])
-                flash('Password has been changed successfully.')
+                flash('Password has been changed successfully.',
+                      category='success')
             return redirect(url_for('profile_bp.display_profile'))
         else:
-            flash('Incorrect current password.')
+            flash('Incorrect current password.', category='error')
             return redirect(url_for('profile_bp.change_password'))
 
 
@@ -75,15 +75,26 @@ def delete_profile_pic(user_id):
 
 
 @login_required
-@profile_bp.route('/v1/profile/delete/<user_id>')
+@profile_bp.route('/v1/profile/delete/<user_id>', methods=['GET', 'POST'])
 def delete_account(user_id):
-    UserDbInter().delete_user(user_id)
-    logout_user()
-    flash('Account deleted successfully')
-    return redirect(url_for('home_bp.index'))
-
-
-@login_required
-@profile_bp.route('/v1/profile/delete/confirm', methods=['GET'])
-def confirm_delete():
-    return render_template('delete_account.html', title='Delete account')
+    if request.method == 'POST':
+        if current_user.oauth_user is False:
+            d = WebInter().get_form_data('password')
+            if not check_password_hash(current_user.password_hash,
+                                       d['password']):
+                flash('Incorrect password.', category='error')
+                return redirect('/v1/profile/delete/{}'.format(user_id))
+            UserDbInter().delete_user(user_id)
+            logout_user()
+            flash('Account deleted successfully.', category='success')
+            return redirect(url_for('home_bp.index'))
+        else:
+            if request.form.get('option') == "True":
+                UserDbInter().delete_user(user_id)
+                logout_user()
+                flash('Account deleted successfully.', category='success')
+                return redirect(url_for('home_bp.index'))
+            else:
+                return redirect('/v1/profile/{}'.format(user_id))
+    else:
+        return render_template('delete_account.html', title='Delete account')
